feat(relay): extract pure enqueueQueueJob core; CLI submitQueueJob wraps it
The hub (embedded in the relay) needs to enqueue background jobs but can't import
apps/cli, and relocating submitQueueJob would cycle (it needs ensureRelay from
sandbox-docker, which depends on relay). Extract the transport-free manifest
builder enqueueQueueJob(input) into @agentbox/relay; submitQueueJob now wraps it
with ensureRelay + the /admin/queue/enqueue poke. Reuses existing loadQueue().

Backlog slice item 3.
