+++ packages/relay/src/index.ts
@@ -161,6 +161,7 @@ export {
   defaultCountRunningBoxes,
   defaultCountWorkingBoxes,
   deleteJob,
+  enqueueQueueJob,
   loadQueue,
   loadQueueConfig,
   occupiesWorkingSlot,
@@ -176,6 +177,8 @@ export {
   waitForFile,
   writeJob,
   type CountWorkingFn,
+  type EnqueueQueueJobInput,
+  type EnqueueQueueJobResult,
   type QueueAgentKind,
   type QueueConfig,
   type QueueJob,
